refactor: sendNotification moved from index
